requirements.txt
numpy==1.26.1
torch==2.4.1
torchvision==0.19.1

pillow==10.4.0
wandb
tqdm